requirements.txt
streamlit
openvino-genai
pypdf
pillow
pytesseract
fpdf2